+++ sandbox/src/Core/Noise.cpp
@@ -6,11 +6,11 @@ namespace Sandbox
 	Noise::Noise() : m_frequency(0.5f)
 	{
 	}
-	float Noise::Get(float x, float y)
+	double Noise::Get(double x, double y)
 	{
 		return m_generator.noise2D_01(x * m_frequency, y * m_frequency);
 	}
-	void Noise::SetFrequency(float frequency)
+	void Noise::SetFrequency(double frequency)
 	{
 		m_frequency = frequency;
 	}
@@ -18,7 +18,7 @@ namespace Sandbox
 	{
 		m_generator.reseed(seed);
 	}
-	float Noise::Noise2D(float x, float y, float frequency)
+	double Noise::Noise2D(double x, double y, double frequency)
 	{
 		static siv::PerlinNoise generator;
 		return generator.noise2D_01(x * frequency, y * frequency);
